✨feat: Add enter key handling to StyledInput component
Adds a new `handle-enter` property to the `StyledInput` component, which
allows the component to handle the Enter key press event. When `handle-enter`
is set to `true`, the component will emit the `enter-pressed` callback when
the Enter key is pressed.

This change enables the `Demo` component to send a message when the user
presses the Enter key while the `StyledInput` component has focus.

Former-commit-id: 74c7232bce6362b1f9bc31c77e18e20bca4b98a7 [formerly 4143ebb874498ec9c4da1e3a1544de9f204538d3] [formerly a503ba4053d6632b6b5ef6173d803b24e7ae0bf6 [formerly 24b67e7361e25feb2125228323794cf4f04a0ac5]]
Former-commit-id: 2a92993517995e73a46994d955e53ab96d2e32b7 [formerly a236d6ef3b342fe4be6375c994b40cb358b76e2f]
Former-commit-id: 3b9c39929579d057bd1f2354f125fa37b48bf56f